Pull demo site webamp config into typed module
